Pop-up modal only fires greater than 768px
Added width and mobile states to track size of screen and whether user is on mobiel. Componentdidmount() fires handleWindowSizeChange() to check width and store in state. handleWindowSizeChange() fires handleWindowSizeCheck to check width state if less than 768 and fires mobile:true if so. Added props.mobile state check to PopupLanding <SyledAlertWrapper> component as part of boolean for popup modal.